--- crates/pop-cli/src/parachains/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod zombienet;
