--- crates/pop-cli/tests/zombienet.toml
@@ -1,24 +1,0 @@
-[relaychain]
-chain = "rococo-local"
-
-[[relaychain.nodes]]
-name = "alice"
-validator = true
-
-[[relaychain.nodes]]
-name = "bob"
-validator = true
-
-[[parachains]]
-id = 1000
-chain = "asset-hub-rococo-local"
-
-[[parachains.collators]]
-name = "asset-hub"
-
-[[parachains]]
-id = 9090
-default_command = "pop-node"
-
-[[parachains.collators]]
-name = "pop"
