+++ crates/pop-parachains/src/lib.rs
@@ -2,10 +2,12 @@ mod build;
 mod generator;
 mod new_pallet;
 mod new_parachain;
+mod up;
 mod utils;
 
 pub use build::build_parachain;
 pub use new_pallet::{create_pallet_template, TemplatePalletConfig};
 pub use new_parachain::{instantiate_template_dir, Config, Template};
-pub use utils::git_helpers::git_init;
+pub use up::Zombienet;
+pub use utils::git::Git;
 pub use utils::pallet_helpers::resolve_pallet_path;
