+++ crates/pop-parachains/src/utils/mod.rs
@@ -1,3 +1,3 @@
-pub mod git_helpers;
+pub mod git;
 pub mod helpers;
 pub mod pallet_helpers;
